MCP Python SDK using Factory (#4)
* MCP Python SDK example using a factory

* Using get_env for consistency

* Using int for consistency

* Updated timestamp

* Use dummy process id

+++ python_examples/ai/mcp/mcp_python_sdk/simple_mcp_server/server.py
@@ -5,6 +5,6 @@ mcp = FastMCP(name='Example MCP Server')
 
 
 @mcp.tool()
-async def add_numbers(a: float, b: float) -> float:
+async def add_numbers(a: int, b: int) -> int:
     """Add two numbers."""
     return a + b
